+++ packages/snap/CHANGELOG.md
@@ -89,7 +89,7 @@
 
 * The following workspace dependencies were updated
   * devDependencies
-    * @safient/snap-types bumped from ^2.1.1 to ^2.1.2
+    * @safient/metasafe-types bumped from ^2.1.1 to ^2.1.2
 
 ## [2.2.0](https://github.com/ChainSafe/filsnap/compare/filsnap-v2.1.0...filsnap-v2.2.0) (2022-04-07)
 
@@ -108,4 +108,4 @@
 
 * The following workspace dependencies were updated
   * devDependencies
-    * @safient/snap-types bumped from ^2.1.0 to ^2.1.1
+    * @safient/metasafe-types bumped from ^2.1.0 to ^2.1.1